Improved server logic & algorithm
