+++ .idea/misc.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectRootManager" version="2" project-jdk-name="Remote Python 3.9.18 Docker (&lt;none&gt;:&lt;none&gt;)" project-jdk-type="Python SDK" />
+</project>
