+++ backend/backend/settings.py
@@ -31,6 +31,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
+    'fhtways',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
