+++ backend/fhtways/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import PathFindingView
+from .views.pathfinding_views import PathFindingView
 
 urlpatterns = [
     path('find-path/', PathFindingView.as_view(), name='find-path'),
